src/legacy/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
librosa>=0.9.0
tqdm>=4.60.0
pyyaml>=6.0
soundfile>=0.12.0
